Remove Node.js CI workflow file
The Node.js CI workflow file (.github/workflows/node.js.yml) has been deleted in this commit. This file was previously responsible for setting up the continuous integration for the Node.js application. The need for this CI workflow setup was evaluated and deemed no longer necessary, thus it was removed.